+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.17.0
+
+* Additional card information, such as prepaid, debit, commercial, Durbin regulated, healthcare, and payroll, are returned on credit card responses
+* Allows transactions to be specified as recurring
+
 ## 2.16.0
 
 * Adds prepaid attribute on credit cards (possible values: Yes, No, Unknown)